refactor: разбивает код скрипта на модули
